// ui/packages/comhairle/src/lib/tools/heyform/types.ts
export type HeyFormFieldKind =
	| 'group'
	| 'welcome'
	| 'thank_you'
	| 'statement'
	| 'short_text'
	| 'long_text'
	| 'number'
	| 'yes_no'
	| 'multiple_choice'
	| 'picture_choice'
	| 'file_upload'
	| 'opinion_scale'
	| 'rating'
	| 'date'
	| 'date_range'
	| 'time'
	| 'input_table'
	| 'payment'
	| 'full_name'
	| 'address'
	| 'email'
	| 'url'
	| 'phone_number'
	| 'country_selector'
	| 'signature'
	| 'legal_terms'
	| 'submit_date'
	| 'hidden_fields'
	| 'variable'
	| 'hidden_checkbox'
	| 'custom_text'
	| 'custom_single'
	| 'custom_multiple'
	| 'custom_date'
	| 'custom_number'
	| 'custom_checkbox';